fix: reorganize between schema & parser
